app/ResearchAgent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ResearchAgent"
version = "0.1.0"
description = "Healthcare research agent on AWS Bedrock AgentCore"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "strands-agents >= 1.13.0",
    "httpx >= 0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
